3B 除了测试2全过. 问题:applych的过程过于慢, 拉低了整体性能. 解决方法, 能否直接通过raft.Start()判断当前节点是否为主.

+++ src/kvraft/client.go
@@ -12,7 +12,7 @@ type Clerk struct {
 	servers []*labrpc.ClientEnd
 	// You will have to modify this struct.
 	master int
-	msgID  int
+	msgID  int64
 	me     int64
 	//sink int
 }
